analyticsMicroservice/requirements.txt
paho-mqtt
Flask==3.0.2
